refactor(SignInDialog): consolidate files
