+++ client/src/dialogs/SignIn/SignInDialog.test.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import { render } from '~/test/helpers/render'
-import SignInDialog from '../SignInDialog'
+import SignInDialog from './SignInDialog'
 
 describe('SignInDialog', () => {
   it('renders', () => {
