+++ client/src/dialogs/SignIn/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SignInDialog'
